Fix crash during upload if connection reset
When any exception occurs during upload, it tries to print the exception to `click.echo`. However, the second argument of `click.echo` is a file that you want to write to. The exception was passed instead, which is not a file, which is was caused the `type object 'ConnectionClosedError' has no attribute 'write')`. As a fix, we now print the exception using normal string formatting.

Closes #23